+++ R/add_constraint_MigClim.R
@@ -38,7 +38,8 @@ NULL
 #' \dontrun{
 #' # Assumes that a trained 'model' object exists
 #'  mod <- scenario(model) |>
-#'   add_predictors(env = predictors, transform = 'scale', derivates = "none") |>
+#'   add_predictors(env = predictors, transform = 'scale',
+#'                  derivates = "none") |>
 #'   add_constraint_MigClim() |>
 #'   project()
 #' }