+++ R/prior_bart.R
@@ -8,7 +8,7 @@ NULL
 #' Bayesian additive regression tree model added via [engine_bart].
 #'
 #' Priors for [engine_bart] have to be specified as transition probabilities of variables which
-#' are interally used to generate splits in the regression tree. Specifying a prior can thus help
+#' are internally used to generate splits in the regression tree. Specifying a prior can thus help
 #' to 'enforce' a split with a given variable.
 #' These can be numeric and coded as values between \code{0} and \code{1}.
 #'